Moved the button struct to it's own file
